requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
atomicwrites==1.4.1
attrs==24.2.0
certifi==2024.8.30
chardet==5.2.0
colorama==0.4.6
configparser==7.1.0
crayons==0.4.0
idna==3.8
importlib-metadata==8.4.0
iniconfig==2.0.0
more-itertools==10.4.0
packaging==24.1
pluggy==1.5.0
py==1.11.0
pyparsing==3.1.4
pytest==8.3.2
requests==2.32.3
selenium==4.24.0
six==1.16.0
toml==0.10.2
urllib3==2.2.2
wcwidth==0.2.13
webdriver-manager==4.0.2
xlrd==2.0.1
zipp==3.20.1
openpyxl~=3.1.5
pandas~=2.2.2